Move store to app/
