+++ src/app/store.ts
@@ -1,13 +1,13 @@
 /** @format */
 
 import { AnyAction, configureStore, ThunkAction } from '@reduxjs/toolkit';
-import { listenerMiddleware } from './app/action-listeners.js';
-import { middleware as dynamicMiddlewares } from './middleware/dynamic-middlewares.js';
-import { buildSliceConfigStoreEnhancer, SliceConfig } from './middleware/slice-configurator.js';
-import { getAppSliceConfigurator } from './slices/app/index.js';
-import { rootReducer, type RootState } from './slices/reducer.js';
+import { middleware as dynamicMiddlewares } from '../middleware/dynamic-middlewares.js';
+import { buildSliceConfigStoreEnhancer, SliceConfig } from '../middleware/slice-configurator.js';
+import { getAppSliceConfigurator } from '../slices/app/index.js';
+import { rootReducer, type RootState } from '../slices/reducer.js';
+import { listenerMiddleware } from './action-listeners.js';
 
-export { RootState } from './slices/reducer.js';
+export { RootState } from '../slices/reducer.js';
 
 export function setupStore(preloadedState?: RootState, hydrate: boolean = true) {
   const sliceConfig: SliceConfig = { disableHydration: !hydrate };
