+++ src/components/lineup-game-live.ts
@@ -6,6 +6,7 @@ import '@material/mwc-icon';
 import { html, LitElement, nothing } from 'lit';
 import { customElement, property, query, state } from 'lit/decorators.js';
 import { map } from 'lit/directives/map.js';
+import { RootState } from '../app/store.js';
 import { ConnectStoreMixin } from '../middleware/connect-mixin.js';
 import { TimerData } from '../models/clock.js';
 import { EventCollectionData } from '../models/events.js';
@@ -45,7 +46,6 @@ import {
   startPeriodCreator,
   toggleClock,
 } from '../slices/live/index.js';
-import { RootState } from '../store.js';
 import './lineup-game-clock.js';
 import { ClockEndPeriodEvent, ClockPeriodData } from './lineup-game-clock.js';
 import './lineup-game-events.js';
