+++ src/components/lineup-view-game-detail.ts
@@ -4,10 +4,10 @@ import '@material/mwc-button';
 import { html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { updateMetadata } from 'pwa-helpers/metadata.js';
+import { RootState } from '../app/store.js';
 import { ConnectStoreMixin } from '../middleware/connect-mixin.js';
 import { GameDetail, GameStatus } from '../models/game.js';
 import { getGame, getGameSliceConfigurator, selectGameById } from '../slices/game/index.js';
-import { RootState } from '../store.js';
 import './lineup-game-complete.js';
 import './lineup-game-live.js';
 import './lineup-game-setup.js';
