+++ src/components/lineup-view-game-roster.ts
@@ -5,6 +5,7 @@ import '@material/mwc-circular-progress';
 import { html, nothing } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { updateMetadata } from 'pwa-helpers/metadata.js';
+import { RootState } from '../app/store.js';
 import { ConnectStoreMixin } from '../middleware/connect-mixin.js';
 import { GameDetail, GameStatus } from '../models/game.js';
 import { Roster } from '../models/player.js';
@@ -16,7 +17,6 @@ import {
   selectGameById,
   selectGameRosterLoading,
 } from '../slices/game/index.js';
-import { RootState } from '../store.js';
 import './lineup-roster.js';
 import { AuthorizedViewElement } from './page-view-element.js';
 import { SharedStyles } from './shared-styles.js';
