+++ src/components/util/auth-controller.ts
@@ -3,7 +3,7 @@
 import { ReactiveController, ReactiveControllerHost } from 'lit';
 import { StateSubscribedController } from '../../middleware/connect-mixin.js';
 import { selectCurrentUserId } from '../../slices/auth/index.js';
-import { RootState } from '../../store.js';
+import { RootState } from '../../app/store.js';
 import { AuthorizedView } from '../page-view-element.js';
 
 type AuthorizedViewHost = ReactiveControllerHost & AuthorizedView;
