+++ src/middleware/slice-configurator.ts
@@ -3,8 +3,8 @@
 import { Reducer, StoreEnhancer, Unsubscribe, nanoid } from '@reduxjs/toolkit';
 import { PersistConfig, persistReducer, persistStore } from 'redux-persist';
 import { AppStartListening, startAppListening } from '../app/action-listeners.js';
+import { AppStore } from '../app/store.js';
 import { rootReducer } from '../slices/reducer.js';
-import { AppStore } from '../store.js';
 import { logger } from '../util/logger.js';
 import { IdbPersistStorage } from './idb-persist-storage.js';
