+++ src/slices/game/roster-logic.ts
@@ -1,11 +1,11 @@
 /** @format */
 
 import { createAsyncThunk, PayloadAction } from '@reduxjs/toolkit';
+import { RootState, ThunkPromise, ThunkResult } from '../../app/store.js';
 import { Player, Roster } from '../../models/player.js';
-import { RootState, ThunkPromise, ThunkResult } from '../../store.js';
 import { loadTeamRoster } from '../team/team-storage.js';
 import { PlayerAddedPayload, RosterCopiedPayload } from './game-action-types.js';
-import { findGame, actions, GameState, selectGameById } from './game-slice.js';
+import { actions, findGame, GameState, selectGameById } from './game-slice.js';
 import { persistGamePlayer } from './game-storage.js';
 
 export const copyRoster = createAsyncThunk<
