+++ src/slices/team/team-storage.ts
@@ -1,12 +1,12 @@
 /** @format */
 
 import { DocumentData } from 'firebase/firestore';
+import { RootState } from '../../app/store.js';
 import { Player, Roster } from '../../models/player.js';
 import { Team, Teams } from '../../models/team.js';
 import { CollectionFilter, reader } from '../../storage/firestore-reader.js';
 import { writer } from '../../storage/firestore-writer.js';
 import { ModelReader } from '../../storage/model-converter.js';
-import { RootState } from '../../store.js';
 import { playerConverter } from '../player/player-storage.js';
 
 const KEY_ROSTER = 'roster';
