+++ src/storage/firestore-writer.ts
@@ -13,11 +13,11 @@ import {
   updateDoc,
   WithFieldValue,
 } from 'firebase/firestore';
+import { RootState } from '../app/store.js';
 import { firebaseRefs } from '../firebase.js';
 import { Model } from '../models/model.js';
 import { selectCurrentTeam } from '../slices/app/index.js';
 import { selectCurrentUserId } from '../slices/auth/index.js';
-import { RootState } from '../store.js';
 import { logger } from '../util/logger.js';
 import { ModelWriter } from './model-converter.js';
