+++ test/unit/helpers/root-state-setup.ts
@@ -1,9 +1,9 @@
 /** @format */
 
+import { RootState } from '@app/app/store.js';
 import { AuthState } from '@app/slices/auth/auth-slice.js';
 import { GameState } from '@app/slices/game/game-slice.js';
 import { LiveState } from '@app/slices/live/index.js';
-import { RootState } from '@app/store.js';
 import sinon from 'sinon';
 import { buildInitialGameState } from './game-state-setup.js';
 import { buildInitialLiveState } from './live-state-setup.js';
